fix: Weekly report end points are using the wrong start date (#48)
* Update to actions/checkout@v3

* fix: computes start date for the weekly reports

Fixes an issue where the weekly report was starting from 7 days ago rather than the beginning of the week

* Allows user to override the start_date

It was too opinionated, restricting the weekly report to only return this weeks data. It now defaults to this weeks data, but the user can override that in the params and return any 7 day period that they want.