feat(tags): adds v9 API tag support (#59)
* feat(tags) Adds v9 support for tags

Updates to use the new endpoints. Must include the `workspace_id` in the call to the tags methods.

* doc(tags): README update

* test(tags): test scaffolidng and create test

* do not the the mapData function

* Adds update tag test

* Adds delete test and TODOs

* Adds delete test and TODOs

* Adds support for empty API responses

the delete endpoints do not return a body and the `client.request` was barfing.

* Updates the unit tests for tags to be idempotent

It will create a tag called `testing-xxx` where xxx is the milliseconds since the epoch. It will update this tag and then delete it, all in one test.

If this test fails, it will leave litter behind.

* skips the tag test

* chore: lintfix

* feat: Adds support for v9 tags endpoints